Implemented Global Exception Handler

+++ src/main/java/eds/cristian/service/impl/TaskServiceImpl.java
@@ -33,6 +33,9 @@ public class TaskServiceImpl implements TaskService {
 
 	@Override
 	public void delete(Long id) {
+		if(!taskRepository.existsById(id)){
+			throw new NoSuchElementException();
+		}
 		taskRepository.deleteById(id);
 		
 	}
